pyproject.toml
[project]
name = "healthcareagent"
version = "0.1.0"
description = "A modular LangGraph agent for healthcare with FHIR API integration and Chainlit UI"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.45",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.5",
    "langchain-core>=0.3.15",
    "chainlit>=1.1.402",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "aiohttp>=3.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
